requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pathlib>=1.0.1
typing-extensions>=4.0.0 